+++ gr-uhd/swig/uhd_swig.i
@@ -54,6 +54,8 @@
 #include <gnuradio/uhd/amsg_source.h>
 %}
 
+%include "gnuradio/uhd/usrp_block.h"
+
 ////////////////////////////////////////////////////////////////////////
 // used types
 ////////////////////////////////////////////////////////////////////////
